requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.1

# Environment Variables
python-dotenv==1.0.0

# Google Gemini AI
google-generativeai==0.3.0

# Optional: Async support
aiofiles==23.2.1

# Development (optional, can be moved to dev-requirements.txt)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.12.0
# ruff==0.1.8
